ch347_repl: Improve user experience with buffers
The read/write commands in the vendor library use the same
buffer for data in and data out, requiring that the buffer
command be called again after each read or write. Instead,
just memcpy the data entered through the buffer command to
a scratchpad buffer that will be passed to the library.